requirements.txt
beautifulsoup4==4.12.3
pillow==10.2.0